Added joint state controller for side joints and ranges are now published on the same topic

+++ pandora_xmega_hardware_interface/include/pandora_xmega_hardware_interface/encoder_sensor.h
@@ -54,7 +54,7 @@ class EncoderSensor : virtual public SensorBase
 
  public:
   
-  float degrees;
+  double degrees;
 };
 
 }  // namespace xmega
